+++ web/app/Controllers/Http/ProfilesController.ts
@@ -1,5 +1,6 @@
 import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import { schema, rules } from '@ioc:Adonis/Core/Validator'
+import Database from '@ioc:Adonis/Lucid/Database'
 
 import User from 'App/Models/User'
 import Profile from 'App/Models/Profile'
@@ -21,8 +22,8 @@ export default class ProfilesController {
         const profile = auth.user.profile[index]
         const profileJSON = profile.toJSON()
 
-        if (profile.chatId !== null) {
-          const user = await User.find(profile.chatId)
+        if (profile.chatUserId !== 0) {
+          const user = await User.find(profile.chatUserId)
           if (user !== null) {
             profileJSON.chat = user.toJSON()
             profiles.push(profileJSON)
@@ -45,7 +46,6 @@ export default class ProfilesController {
 
       const idNumber = Number(id)
       if (Number.isNaN(idNumber)) {
-      // Make sure it's a number.
         session.flash('message', { error: 'Error: Parameter is not a number. Profile url can only contain numbers.' })
         return response.redirect('/')
       }
@@ -60,8 +60,8 @@ export default class ProfilesController {
       const ownProfile = profile.userId === authID
 
       let chat: User | null = null
-      if (profile.chatId !== null && !profile.global) {
-        chat = await User.find(profile.chatId)
+      if (profile.chatUserId !== 0) {
+        chat = await User.find(profile.chatUserId)
         // Should not happen, ever.
         if (chat === null) {
           session.flash('message', { error: 'Error: Chat owner is missing. Report this error on github.' })
@@ -69,11 +69,26 @@ export default class ProfilesController {
         }
       }
 
-      await profile.preload('matches')
       await auth.user.preload('favoriteStreamers')
 
       const userJSON = auth.user.toJSON()
       const profileJSON = profile.toJSON()
+
+      const matches = await profile.related('matches').query().limit(10)
+      if (matches !== null) {
+        profileJSON.matches = []
+        for (let index = 0; index < matches.length; index++) {
+          const profile = matches[index]
+          const user = await User.find(profile.userId)
+
+          if (user !== null) {
+            profileJSON.matches.push({
+              profile: profile.toJSON(),
+              user: user.toJSON(),
+            })
+          }
+        }
+      }
 
       if (ownProfile) {
       // If own, allow access.
@@ -87,20 +102,21 @@ export default class ProfilesController {
           },
         })
       } else {
-        // If someone else's, check if they've matched for that specific profile.
-        const matched = profile.matches.find(match => match.id === authID)
-        if (matched !== null) {
+        // If someone else's, check if this user has matched for that specific profile.
+        const match = await Database.query().from('matches_lists').where({
+          profile_id: profile.id,
+          match_user_id: authID,
+        }).first()
+
+        if (match !== null) {
           const userOfProfile = await User.find(profile.userId)
           if (userOfProfile !== null) {
-            delete profile.matches // We don't want to send that user's matches.
-
             await userOfProfile.preload('favoriteStreamers')
 
             return view.render('core', {
               user: userJSON,
               profile: profileJSON,
               profileUser: userOfProfile.toJSON(),
-              guest: true, // Required.
               web: {
                 template: 'profile',
                 title: chat !== null
@@ -110,7 +126,7 @@ export default class ProfilesController {
             })
           }
         }
-        // Disallow access.
+
         session.flash('message', { error: 'Error: You are not allowed access to that profile!' })
         return response.redirect('/')
       }
@@ -126,7 +142,6 @@ export default class ProfilesController {
 
     const idNumber = Number(id)
     if (Number.isNaN(idNumber)) {
-      // Make sure it's a number.
       session.flash('message', { error: 'Error: Parameter is not a number. Profile url can only contain numbers.' })
       return response.redirect('/profile/')
     }
@@ -163,7 +178,6 @@ export default class ProfilesController {
       session.flash('message', { message: 'Successfully updated your profile.' })
       return response.redirect(`/profile/${id}`)
     } else {
-      // Disallow access.
       session.flash('message', { error: 'Error: You are not allowed access to that profile!' })
       return response.redirect('/profile/')
     }
@@ -180,15 +194,70 @@ export default class ProfilesController {
 
     const profile = auth.user.profile.find(profile => profile.id === id)
     if (profile !== undefined) {
-      await profile.preload('matches')
-
       await profile.related('matches').detach()
+
+      await Database.query().from('matches_lists').where('match_profile_id', profile.id).delete()
 
       await profile.delete()
       session.flash('message', { message: 'Profile has been deleted.' })
     }
 
     return response.redirect('/profile/')
+  }
+
+  public async matches ({ params, request, auth, response }: HttpContextContract) {
+    if (auth.user === undefined) {
+      return
+    }
+
+    const method = request.intended()
+    if (method !== 'POST') {
+      return
+    }
+
+    const { id } = params
+    const { pagination } = request.get()
+
+    const idNumber = Number(id)
+    if (Number.isNaN(idNumber)) {
+      return response.badRequest({ error: 'Error: Parameter is not a number. Profile url can only contain numbers.' })
+    }
+
+    const paginationNumber = Number(pagination)
+    if (Number.isNaN(paginationNumber)) {
+      return response.badRequest({ error: 'Error: Pagination get() query is not a number.' })
+    }
+
+    const profile = await Profile.find(idNumber)
+
+    if (profile === null) {
+      return response.notFound({ error: 'Error: Profile does not exist.' })
+    }
+
+    const ownProfile = profile.userId === auth.user.id
+
+    if (ownProfile) {
+      const matches = await profile.related('matches').query().offset(paginationNumber).limit(10)
+      const matchesJSON: any[] = []
+
+      if (matches !== null) {
+        for (let index = 0; index < matches.length; index++) {
+          const profile = matches[index]
+          const user = await User.find(profile.userId)
+
+          if (user !== null) {
+            matchesJSON.push({
+              profile: profile.toJSON(),
+              user: user.toJSON(),
+            })
+          }
+        }
+      }
+
+      return response.json(matchesJSON)
+    } else {
+      return response.forbidden({ error: 'Error: You are not allowed access to that profile!' })
+    }
   }
 
   public async unmatch ({ params, request, auth, response }: HttpContextContract) {
@@ -202,41 +271,50 @@ export default class ProfilesController {
     }
 
     const { id } = params
-    const { userId } = request.post()
+    const { profileID } = request.post()
 
-    if (userId === undefined) {
-      return response.badRequest({ error: 'Body must contain { userId: number }.' })
+    if (profileID === undefined) {
+      return response.badRequest({ error: 'Error: Body must contain { profileID: number }.' })
+    }
+
+    const profileIDNumber = Number(profileID)
+    if (Number.isNaN(profileIDNumber)) {
+      return response.badRequest({ error: 'Error: profileID is not a number.' })
     }
 
     const idNumber = Number(id)
     if (Number.isNaN(idNumber)) {
-      // Make sure it's a number.
-      return response.badRequest({ error: 'Parameter is not a number. Profile url can only contain numbers.' })
+      return response.badRequest({ error: 'Error: Parameter is not a number. Profile url can only contain numbers.' })
     }
 
     const profile = await Profile.find(idNumber)
 
     if (profile === null) {
-      return response.notFound({ error: 'Profile does not exist.' })
+      return response.notFound({ error: 'Error: Profile does not exist.' })
     }
 
     const ownProfile = profile.userId === auth.user.id
 
     if (ownProfile) {
-      await profile.preload('matches')
+      const match = await profile.related('matches').query().where('id', profileIDNumber).first()
 
-      const match = profile.matches.find(user => user.id === userId)
-
-      if (match !== undefined) {
-        await profile.related('matches').detach([match.id])
-
-        return response.ok({ success: true })
-      } else {
-        return response.notFound({ error: 'Could not remove. That match does not exist.' })
+      if (match === null) {
+        return response.notFound({ error: 'Error: Could not remove. That matched profile does not exist.' })
       }
+
+      // Remove match from this profile.
+      await profile.related('matches').detach([match.id])
+
+      // Remove match from match's profile
+      await Database.from('matches_lists').where({
+        profile_id: profileIDNumber,
+        match_user_id: auth.user.id,
+        match_profile_id: profile.id,
+      }).delete()
+
+      return response.ok({ message: 'Successfully removed match from profile.' })
     } else {
-      // Disallow access.
-      return response.forbidden({ error: 'You are not allowed access to that profile!' })
+      return response.forbidden({ error: 'Error: You are not allowed access to that profile!' })
     }
   }
 }