+++ web/database/migrations/1_favorite_streamer_lists.ts
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/no-floating-promises */
+import { DateTime } from 'luxon'
 import BaseSchema from '@ioc:Adonis/Lucid/Schema'
 
 export default class FavoriteStreamerLists extends BaseSchema {
@@ -8,7 +9,8 @@ export default class FavoriteStreamerLists extends BaseSchema {
     this.schema.createTable(this.tableName, table => {
       table.integer('user_id')
       table.integer('streamer_id')
-      table.timestamps(true)
+      table.dateTime('created_at').defaultTo(DateTime.fromJSDate(new Date()).toFormat('yyyy-MM-dd HH:mm:ss'))
+      table.dateTime('updated_at').defaultTo(DateTime.fromJSDate(new Date()).toFormat('yyyy-MM-dd HH:mm:ss'))
     })
   }