+++ web/database/migrations/1_matches_lists.ts
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/no-floating-promises */
+import { DateTime } from 'luxon'
 import BaseSchema from '@ioc:Adonis/Lucid/Schema'
 
 export default class MatchesLists extends BaseSchema {
@@ -6,9 +7,12 @@ export default class MatchesLists extends BaseSchema {
 
   public async up () {
     this.schema.createTable(this.tableName, table => {
+      table.integer('profile_id')
       table.integer('user_id')
-      table.integer('match_id')
-      table.timestamps(true)
+      table.integer('match_profile_id')
+      table.integer('match_user_id')
+      table.dateTime('created_at').defaultTo(DateTime.fromJSDate(new Date()).toFormat('yyyy-MM-dd HH:mm:ss'))
+      table.dateTime('updated_at').defaultTo(DateTime.fromJSDate(new Date()).toFormat('yyyy-MM-dd HH:mm:ss'))
     })
   }